move critically damped blender to data_blending to be available for use outside of stock extrapolation
